Moving this into its own type
