+++ ImportFixer/KAImportType.h
@@ -1,0 +1,14 @@
+//
+//  KAImportType.h
+//  ImportFixer
+//
+//  Created by Kenneth Parker Ackerson on 1/30/16.
+//  Copyright © 2016 Kenneth Parker Ackerson. All rights reserved.
+//
+
+typedef NS_ENUM(char, KAImportType) {
+    KAImportTypePound,
+    KAImportTypePoundLibrary,
+    KAImportTypeAtSign,
+    KAImportTypeSwift
+};
